pyproject.toml
[tool.poetry]
name = "apolo11-project"
version = "0.1.0"
description = ""
authors = ["Brahian2016 <bram16.bam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
